Put rnnlm in examples to clean the root directory
