+++ examples/rnnlm/install-examples.sh
