--- rnnlm/CMakeLists.txt
@@ -1,15 +1,0 @@
-CMAKE_MINIMUM_REQUIRED(VERSION 2.8)
-
-foreach(TARGET lm)
-  ADD_EXECUTABLE(${TARGET} ${TARGET}.cc)
-  target_link_libraries(${TARGET} dynet ${LIBS})
-  if(UNIX AND NOT APPLE)
-    target_link_libraries(${TARGET} rt)
-  endif()
-  if (WITH_CUDA_BACKEND)
-    add_dependencies(${TARGET} dynetcuda)
-    target_link_libraries(${TARGET} dynetcuda)
-    CUDA_ADD_CUBLAS_TO_TARGET(${TARGET})
-  endif (WITH_CUDA_BACKEND)
-endforeach()
-
